Edit: Algoritme.java set afstand to float Edit: Resultaten.java Added afstand Edit: ResultatenVergelijking.java Add best afstand Edit: Simulatie.java set bigger tabbedpane size Edit: SimulatiePanel.java add labels and layouts Edit: WillekeurigAlgoritme.java added stand calculate

+++ TSP/src/me/TSP/Algoritme.java
@@ -10,7 +10,7 @@ public abstract class Algoritme
     private Long time;
     private String name;
     private int simulatieNr;
-    private int aftstand;
+    private float aftstand;
     private ArrayList<Vak> bestOrderLocaties = new ArrayList<>();
 
     public void calculate(ArrayList<Vak> locaties) {}
@@ -46,11 +46,11 @@ public abstract class Algoritme
         this.bestOrderLocaties = a;
     }
 
-    public int getAftstand() {
+    public float getAftstand() {
         return aftstand;
     }
 
-    public void setAftstand(int aftstand) {
+    public void setAftstand(float aftstand) {
         this.aftstand = aftstand;
     }
 }
